products api

+++ users/models.py
@@ -4,11 +4,8 @@ from django.contrib.auth.models import AbstractUser
 
 class User(AbstractUser):
     phone = models.CharField(max_length=15, null=True)
-    age = models.PositiveIntegerField(null=True)
+    age = models.PositiveSmallIntegerField(null=True)
     address = models.CharField(max_length=255, null=True)
 
     def __str__(self):
         return self.username
-
-
-
